Fault reset button and public booleans
Added a reset button to clear the latched faults if needed. This code also logs if the button is pressed so that nothing is truly hidden from us. Also this commit captured the addition of public booleans from the monitor. These values will be what other subsystems use to act on this information.